Pagination, take 1.
This involved a fair bit of re-work in Krang::ElementClass->fill_template(), which is heavily documented in-line.  It works, so far it seems to be as fast as the previous code (which had no pagination), but it's somewhat unwieldy - further work is needed.

Some pagination testing exists, more rigerous tests of multi-page articles are still needed.


git-svn-id: http://svn.krangcms.com/trunk/krang@1400 689d98c0-9023-0410-9c0b-f79783857903