Preview Windows are now targeted to the instance.
What this means is that if you leave a preview window open, subsuquent previews of stories/media in the same instance will reload in that window.

The window will only be brought to the front if your Mozilla/other browser allows Javascript to do so (annoying debug gotcha that I hit).

Implemented in the WorkSpace, Story, and Media interfaces.  I think that's all the places one can preview.


git-svn-id: http://svn.krangcms.com/trunk/krang@1452 689d98c0-9023-0410-9c0b-f79783857903