Moved Krang user/group creation from krang_install to Krang::Platform. This can now be overridden on a platform-by-platform basis.
Created Krang::Platform->find_bin().  Use of this sub now needs to be
propegated throughout krang_install along with Krang::Platform and its
derivatives - it searches $PATH for the full path for the executable
you're trying to run.

platform/Redhat9/Platform.pm is the first to use the new find_bin() code.


git-svn-id: http://svn.krangcms.com/trunk/krang@2418 689d98c0-9023-0410-9c0b-f79783857903