changing the category-link triggered publishes to be faster Before this change, each publish looked at every element in every story to see if there were any category links with the publish_if_modified_X links in them. This was drastically slow on any system with a large number of stories.
Now we save entries into a new story_category_link table when stories
are saved and they have category links and then we use that table
to get the list of possible storys to add to the publish.

Also changed the code to not have the publish be conditional on the
story actually being changed (the is_modified() stuff). This is because
we still want the category-linked story to publish even if the target
story hasn't changed in case there are changes in element libs or templates
or media files, etc that should be picked up as well.



git-svn-id: http://svn.krangcms.com/trunk/krang@6735 689d98c0-9023-0410-9c0b-f79783857903